fixing unconnected draggable tests
